Move parse_prior_string to distributions.py, as well
(And change its layout, while we're changing those lines anyway.)
