+++ beastling/configuration.py
@@ -18,6 +18,7 @@ from clldutils.dsv import reader
 from clldutils.path import Path
 
 from beastling.fileio.datareaders import load_location_data
+from .distributions import parse_prior_string
 import beastling.clocks.strict as strict
 import beastling.clocks.relaxed as relaxed
 import beastling.clocks.random as random_clock
@@ -1109,133 +1110,9 @@ class Configuration(object):
                 self.calibrations[clade_identifier] = cal_obj
 
     def parse_calibration_string(self, orig_clade, cs, is_tip_cal=False):
-        return self.parse_prior_string(
+        return parse_prior_string(
             prior_name="calibration of clade {:}".format(orig_clade),
             cs=cs, is_point=is_tip_cal)
-
-    @staticmethod
-    def parse_prior_string(cs, prior_name="?", is_point=False):
-        """Parse a prior-describing string.
-
-        The basic format of such a string is [offset + ]distribution
-        Offset is a number, distribution can describe a probability
-        density function of normal, lognormal (including rlognormal, a
-        reparametrization where the mean is given in real space, not
-        in log space) or uniform type in one of the following
-        ways. Pseudo-densities with infinite interval are permitted.
-
-        >>> parse = Configuration.parse_prior_string
-        >>> # Parameters of a normal distribution
-        >>> parse("0, 1")
-        (0.0, 'normal', (0.0, 1.0))
-        >>> # Parameters of some other distribution
-        >>> parse(" rlognormal(1, 1)")
-        (0.0, 'lognormal', (0.0, 1.0))
-        >>> # A distribution shape and its 95%-interval
-        >>> parse("normal (1-5)")
-        (0.0, 'normal', (3.0, 1.0204081632653061))
-        >>> parse("1 - 5")
-        (0.0, 'normal', (3.0, 1.0204081632653061))
-        >>> parse(">1200")
-        (0.0, 'uniform', (1200.0, 9223372036854775807))
-        >>> parse("< 1200")
-        (0.0, 'uniform', (0.0, 1200.0))
-
-        ====
-        Note
-        ====
-
-        For uniform distributions, "uniform(0-1) does not give the 95%
-        interval, but the lower and upper bounds.
-        
-        While the ">1" and "<1" notations generate uniform
-        distributions, the bare "0-1" notation generates a normal
-        distribution.
-
-        =======
-        Returns
-        =======
-        offset: int
-        type: str
-            A known distribution type
-        parameters: tuple of floats
-            The parameters for that distribution
-
-        """
-        orig_cs = cs[:]
-        # Find offset
-        if cs.count("+") == 1:
-            os, dist = cs.split("+")
-            offset = float(os.strip())
-            cs = dist.strip()
-        else:
-            offset = 0.0
-
-        # Parse distribution
-        if cs.count("(") == 1 and cs.count(")") == 1:
-            dist_type, cs = cs.split("(", 1)
-            dist_type = dist_type.strip().lower()
-            if dist_type not in ("uniform", "normal", "lognormal", "rlognormal"):
-                raise ValueError("Prior Specification \"%s\" for %s uses an unknown distribution %s!" % (orig_cs, prior_name, dist_type))
-            cs = cs[0:-1]
-        else:
-            # Default to normal
-            dist_type = "normal"
-
-        # Parse / infer params
-        if cs.count(",") == 1 and not any([x in cs for x in ("<", ">")]):
-            # We've got explicit params
-            p1, p2 = map(float,cs.split(","))
-        elif cs.count("-") == 1 and not any([x in cs for x in (",", "<", ">")]):
-            # We've got a 95% HPD range
-            lower, upper = map(float, cs.split("-"))
-            if upper <= lower:
-                raise ValueError("Prior Specification \"%s\" for %s has an upper bound which is not higher than its lower bound!" % (orig_cs, prior_name))
-            mid = (lower+upper) / 2.0
-            if dist_type == "normal":
-                p1 = (upper + lower) / 2.0
-                p2 = (upper - mid) / 1.96
-            elif dist_type == "lognormal":
-                p1 = math.log(mid)
-                p2a = (p1 - math.log(lower)) / 1.96
-                p2b = (math.log(upper) - p1) / 1.96
-                p2 = (p2a+p2b)/2.0
-            elif dist_type == "uniform":
-                p1 = lower
-                p2 = upper
-        elif (cs.count("<") == 1 or cs.count(">") == 1) and not any([x in cs for x in (",", "-")]):
-            # We've got a single bound
-            dist_type = "uniform"
-            sign, bound = cs[0], cs[1:]
-            if sign == "<":
-                p1 = 0.0
-                p2 = float(bound.strip())
-            elif sign == ">":
-                p1 = float(bound.strip())
-                p2 = sys.maxsize
-            else:
-                raise ValueError("Could not parse prior specification \"%s\" for %s" % (orig_cs, prior_name))
-        elif is_point:
-            # Last chance: It's a single language pinned to a
-            # single date, so make sure to pin it to that date
-            # late and nothing else is left to do with this
-            # prior specification.
-            try:
-                dist_type = "point"
-                p1 = float(cs)
-                p2 = p1
-            except ValueError:
-                raise ValueError("Could not parse prior specification \"%s\" for %s" % (orig_cs, prior_name))
-        else:
-            raise ValueError("Could not parse prior specification \"%s\" for %s" % (orig_cs, prior_name))
-
-        # If this is a lognormal prior specification with the mean in realspace, adjust
-        if dist_type == "rlognormal":
-            p1 = math.log(p1)
-            dist_type = "lognormal"
-
-        # All done!
-        return offset, dist_type, (p1, p2)
 
     def get_languages_by_glottolog_clade(self, clade):
         """
